Removed ChatService Interface, renamed AtherysChatService to ChatService
